fix fixed point intrinsics and some compiler warnings (#3593)
* fix fixed point intrinsics and some compiler warnings

* restore signed saturate asm and float conversion code

* fix rounding again

* Fix rounding

* fix name for clang too

---------

+++ src/deluge/dsp/oscillators/oscillator.cpp
@@ -274,11 +274,11 @@ doSaw:
 
 				if (!doOscSync) {
 					if (applyAmplitude) {
-						dsp::renderCrudeSawWave({bufferStart, numSamples}, phase, phaseIncrement, amplitude,
-						                        amplitudeIncrement);
+						dsp::renderCrudeSawWave({bufferStart, static_cast<uint>(numSamples)}, phase, phaseIncrement,
+						                        amplitude, amplitudeIncrement);
 					}
 					else {
-						dsp::renderCrudeSawWave({bufferStart, numSamples}, phase, phaseIncrement);
+						dsp::renderCrudeSawWave({bufferStart, static_cast<uint>(numSamples)}, phase, phaseIncrement);
 					}
 					return;
 				}
